+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Released on Friday, July 31 2026
 
 - Improved CSS selector performance (#1270, #929) @lahma
+- Added support for the `:host` pseudo selector (#1271)
 - Added support for the DomGeometry spec (#1193)
 
 # 1.6.0